requirements.txt
# pip install -r requirements.txt
pycparser
jmcomic
pillow
pyyaml
img2pdf